refactoring: MainController 분리
1. MainController 내 Notification 관련 메소드들 NotificationController로 이동

Resolves:
Ref:
Related to:
